+++ component/backend/Toolbar/Toolbar.php
@@ -39,7 +39,6 @@ class Toolbar extends \FOF30\Toolbar\Toolbar
 			'COM_AKEEBASUBS_MAINMENU_INVOICES' => [
 				'Invoices',
 				'CreditNotes',
-				'CreditNoteTemplates',
 			],
 		];
 
@@ -124,25 +123,6 @@ class Toolbar extends \FOF30\Toolbar\Toolbar
 
 			$this->addCustomBtn('test-template', $options);
 		}
-	}
-
-	public function onImportsDefault()
-	{
-		$this->renderSubmenu();
-
-		$option = $this->container->componentName;
-		$view = 'Import';
-
-		$subtitle_key = $option . '_TITLE_' . $view;
-		JToolBarHelper::title(JText::_($option).' &ndash; <small>' .
-		                      JText::_($subtitle_key) .
-		                      '</small>',
-			str_replace('com_', '', $option));
-
-		JToolBarHelper::custom('import', 'download', 'download', 'COM_AKEEBASUBS_IMPORT', false);
-		JToolbarHelper::divider();
-
-		JToolBarHelper::back('JTOOLBAR_BACK', 'index.php?option=com_akeebasubs&view=ControlPanel');
 	}
 
 	public function onTaxconfigsMain()