+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 * Removed post-installation wizard (gh-335)
 * Removed the Import feature (gh-340)
 * Removed the API Coupons feature (gh-352)
+* Removed the Invoice Templates feature (gh-345)
+* Removed the Credit Notes Templates feature (gh-347)
 
 # 6.2.3